Exclude launch tests from integration tests
