+++ test/optimism/_launch.test.ts
@@ -1,15 +1,15 @@
 import { assert } from "chai";
 
 import env from "../../utils/env";
-import arbitrum from "../../utils/arbitrum";
-import { L1ERC20TokenBridge__factory } from "../../typechain";
 import { wei } from "../../utils/wei";
+import optimism from "../../utils/optimism";
 import testing, { scenario } from "../../utils/testing";
 import { BridgingManagerRole } from "../../utils/bridging-management";
+import { L1ERC20TokenBridge__factory } from "../../typechain";
 
 const REVERT = env.bool("REVERT", true);
 
-scenario("Arbitrum :: Launch integration test", ctx)
+scenario("Optimism :: Launch integration test", ctxFactory)
   .after(async (ctx) => {
     if (REVERT) {
       await ctx.l1Provider.send("evm_revert", [ctx.snapshot.l1]);
@@ -22,28 +22,28 @@ scenario("Arbitrum :: Launch integration test", ctx)
   })
 
   .step("Enable deposits", async (ctx) => {
-    const { l1ERC20TokenGateway } = ctx;
-    assert.isFalse(await l1ERC20TokenGateway.isDepositsEnabled());
+    const { l1ERC20TokenBridge } = ctx;
+    assert.isFalse(await l1ERC20TokenBridge.isDepositsEnabled());
 
-    await l1ERC20TokenGateway.enableDeposits();
-    assert.isTrue(await l1ERC20TokenGateway.isDepositsEnabled());
+    await l1ERC20TokenBridge.enableDeposits();
+    assert.isTrue(await l1ERC20TokenBridge.isDepositsEnabled());
   })
 
   .step("Renounce role", async (ctx) => {
-    const { l1ERC20TokenGateway, l1DevMultisig } = ctx;
+    const { l1ERC20TokenBridge, l1DevMultisig } = ctx;
     assert.isTrue(
-      await l1ERC20TokenGateway.hasRole(
+      await l1ERC20TokenBridge.hasRole(
         BridgingManagerRole.DEPOSITS_ENABLER_ROLE.hash,
         await l1DevMultisig.getAddress()
       )
     );
 
-    await l1ERC20TokenGateway.renounceRole(
+    await l1ERC20TokenBridge.renounceRole(
       BridgingManagerRole.DEPOSITS_ENABLER_ROLE.hash,
       await l1DevMultisig.getAddress()
     );
     assert.isFalse(
-      await l1ERC20TokenGateway.hasRole(
+      await l1ERC20TokenBridge.hasRole(
         BridgingManagerRole.DEPOSITS_ENABLER_ROLE.hash,
         await l1DevMultisig.getAddress()
       )
@@ -52,10 +52,17 @@ scenario("Arbitrum :: Launch integration test", ctx)
 
   .run();
 
-async function ctx() {
-  const networkName = env.network("TESTING_ARB_NETWORK", "mainnet");
-  const { l1Provider, l2Provider, l1ERC20TokenGateway, l1DevMultisig } =
-    await arbitrum.testing(networkName).getIntegrationTestSetup();
+async function ctxFactory() {
+  const networkName = env.network("TESTING_OPT_NETWORK", "mainnet");
+
+  const { l1Provider, l2Provider, l1ERC20TokenBridge } = await optimism
+    .testing(networkName)
+    .getIntegrationTestSetup();
+
+  const hasDeployedContracts = testing.env.USE_DEPLOYED_CONTRACTS(false);
+  const l1DevMultisig = hasDeployedContracts
+    ? await testing.impersonate(testing.env.L1_DEV_MULTISIG(), l1Provider)
+    : testing.accounts.deployer(l1Provider);
 
   const l1Snapshot = await l1Provider.send("evm_snapshot", []);
   const l2Snapshot = await l2Provider.send("evm_snapshot", []);
@@ -63,12 +70,12 @@ async function ctx() {
   const l1Sender = testing.accounts.sender(l1Provider);
 
   await l1Sender.sendTransaction({
-    value: wei`1 ether`,
     to: await l1DevMultisig.getAddress(),
+    value: wei.toBigNumber(wei`1 ether`),
   });
 
-  const l1ERC20TokenGatewayImpl = L1ERC20TokenBridge__factory.connect(
-    l1ERC20TokenGateway.address,
+  const l1ERC20TokenBridgeImpl = L1ERC20TokenBridge__factory.connect(
+    l1ERC20TokenBridge.address,
     l1DevMultisig
   );
 
@@ -76,7 +83,7 @@ async function ctx() {
     l1Provider,
     l2Provider,
     l1DevMultisig,
-    l1ERC20TokenGateway: l1ERC20TokenGatewayImpl,
+    l1ERC20TokenBridge: l1ERC20TokenBridgeImpl,
     snapshot: {
       l1: l1Snapshot,
       l2: l2Snapshot,
